requirements.txt
streamlit
langchain
langchain_community
langchain_groq
python_dotenv
duckduckgo_search
